+++ src/multiLine/index.ts
@@ -4,6 +4,7 @@ import type {AttributesToken, ExtToken} from '../../internal';
 /* NOT FOR BROWSER */
 
 import {classes} from '../../util/constants';
+import {clone} from '../../mixin/clone';
 
 /* NOT FOR BROWSER END */
 
@@ -49,6 +50,14 @@ export abstract class MultiLineToken extends Token {
 	/** @private */
 	override print(): string {
 		PRINT: return super.print({sep: '\n'});
+	}
+
+	/* NOT FOR BROWSER */
+
+	@clone
+	override cloneNode(): this {
+		const C = this.constructor as new (...args: any[]) => this;
+		return new C(undefined, this.getAttribute('config'));
 	}
 }