+++ src/multiLine/inputbox.ts
@@ -12,8 +12,8 @@ import {classes} from '../../util/constants';
 
 /** `<inputbox>` */
 export abstract class InputboxToken extends ParamTagToken {
-	/** @class */
-	constructor(include: boolean, wikitext: string | undefined, config: Config, accum: Token[] = []) {
+	/** @param name 扩展标签名 */
+	constructor(name: string, include: boolean, wikitext: string | undefined, config: Config, accum: Token[] = []) {
 		const placeholder = Symbol('InputboxToken'),
 			newConfig = config.excludes.includes('heading')
 				? config
@@ -26,7 +26,7 @@ export abstract class InputboxToken extends ParamTagToken {
 		wikitext &&= parseCommentAndExt(wikitext, newConfig, accum, include);
 		wikitext &&= parseBraces(wikitext, newConfig, accum);
 		accum.splice(length, 1);
-		super(include, wikitext, newConfig, accum, {
+		super(name, include, wikitext, newConfig, accum, {
 			ArgToken: ':', TranscludeToken: ':',
 		});
 	}