+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Added**
 
 - [`AttributesToken.prototype.css`](https://github.com/bhsd-harry/wikiparser-node/wiki/AttributesToken-%28EN%29#css) now inserts the `style` attribute if not present
+- [`ParamLineToken.prototype.lint`](https://github.com/bhsd-harry/wikiparser-node/wiki/ParamLineToken-%28EN%29#lint) now reports unknown parameter names
 
 **Fixed**
 
@@ -13,7 +14,6 @@
 - [`AttributesToken.classList`](https://github.com/bhsd-harry/wikiparser-node/wiki/AttributesToken-%28EN%29#classlist) when the `class` attribute contains multiple spaces between class names
 - [`AttributesToken.prototype.sanitize`](https://github.com/bhsd-harry/wikiparser-node/wiki/AttributesToken-%28EN%29#sanitize) should leave spaces between attributes
 - [`Token.prototype.createElement`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token-%28EN%29#createelement) for extension tags
-- [`ParamTagToken.prototype.lint`](https://github.com/bhsd-harry/wikiparser-node/wiki/ParamTagToken-%28EN%29#lint) should ignore spaces at SOL
 
 ## v1.32.0